Remove unnecessary files and non-functional pages/components
