+++ client/src/App.jsx
@@ -8,10 +8,7 @@ import {
     Solo,
     Match,
     MatchEnd,
-    Matchmaking,
     MatchSetup,
-    Arcaneum,
-    Collection,
     Deck,
     Packs,
     FirstDeck,
@@ -38,9 +35,6 @@ function App() {
                         <Route path='/matchSetup' element={<MatchSetup />} />
                         <Route path='/match' element={<Match />} />
                         <Route path='/matchEnd' element={<MatchEnd />} />
-                        <Route path='/matchmaking' element={<Matchmaking />} />
-                        <Route path='/arcaneum' element={<Arcaneum />} />
-                        <Route path='/collection' element={<Collection />} />
                         <Route path='/deck' element={<Deck />} />
                         <Route path='/packs' element={<Packs />} />
                         <Route path='/account' element={<Account />} />
