+++ client/src/assets/icons/index.js
@@ -1,7 +1,4 @@
 import coin from './coin.png'
-import gift from './gift.png'
 import home from './home.png'
-import returnArrow from './return.png'
-import settings from './settings.png'
 
-export { coin, gift, home, returnArrow, settings }
+export { coin, home }
