+++ client/src/assets/misc/index.js
@@ -1,7 +1,3 @@
-import creature1 from './creature1.png'
-import creature2 from './creature2.png'
-import creature3 from './creature3.png'
 import defaultPlayer from './defaultPlayer.png'
-import spellCircle from './spellcircle.png'
 
-export { creature1, creature2, creature3, defaultPlayer, spellCircle }
+export { defaultPlayer }
