+++ client/src/assets/packs/index.js
@@ -1,5 +1,5 @@
-import largePack from './pack3v2.png'
-import mediumPack from './pack2v2.png'
-import smallPack from './pack1v2.png'
+import largePack from './pack3.png'
+import mediumPack from './pack2.png'
+import smallPack from './pack1.png'
 
 export { largePack, mediumPack, smallPack }
