+++ client/src/assets/pageImages/index.js
@@ -1,31 +1,9 @@
 import city from './city.jpg'
 import crystal from './crystal.jpg'
-import darkportal from './darkportal.jpg'
-import darkskies from './darkskies.jpg'
 import dungeon from './shuxingli.jpg'
-import entrance from './entrance.jpg'
 import facecave from './facecave.jpg'
 import floating from './floating.jpg'
 import library1 from './library1.jpg'
-import library2 from './library2.jpg'
 import snakecave from './snakecave.jpg'
-import snowcastle from './snowcastle.jpg'
-import treehouse from './treehouse.jpg'
-import view from './view.jpg'
 
-export {
-	city,
-	crystal,
-	darkportal,
-	darkskies,
-	dungeon,
-	entrance,
-	facecave,
-	floating,
-	library1,
-	library2,
-	snakecave,
-	snowcastle,
-	treehouse,
-	view,
-}
+export { city, crystal, dungeon, facecave, floating, library1, snakecave }
