--- client/src/assets/ranks/index.js
@@ -1,4 +1,0 @@
-import rank1 from './rank1.png'
-import rank6 from './rank6.png'
-
-export { rank1, rank6 }
