+++ client/src/components/Card/Card.scss
@@ -63,11 +63,11 @@
     }
 
     &.blue {
-        background-image: url('../../assets/backgrounds/blue1.jpg');
+        background-image: url('../../assets/backgrounds/blue.jpg');
     }
 
     &.red {
-        background-image: url('../../assets/backgrounds/red1.jpg');
+        background-image: url('../../assets/backgrounds/red.jpg');
     }
 
     &.transparent {
