+++ client/src/components/index.js
@@ -3,4 +3,3 @@ export { default as Button } from './Button/Button'
 export { default as Card } from './Card/Card'
 export { default as Cell } from './Cell/Cell'
 export { default as TextInput } from './TextInput/TextInput'
-export { default as TurnArrow } from './TurnArrow/TurnArrow'
