--- client/src/pages/Arcaneum/Arcaneum.scss
@@ -1,117 +1,0 @@
-.arcaneum {
-    align-items: center;
-    justify-content: center;
-    background-image: url('../../assets/pageImages/entrance.jpg');
-    background-position: center;
-    background-size: cover;
-
-    .container {
-        position: relative;
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        gap: 48px;
-        padding-top: 80px;
-
-        @media screen and (max-width: 900px) {
-            padding: 120px 0 40px 0;
-        }
-
-        .top,
-        .middle,
-        .bottom {
-            display: flex;
-
-            @media screen and (max-width: 900px) {
-                width: 75vw;
-                align-items: center;
-                justify-content: space-between;
-            }
-
-            @media screen and (max-width: 600px) {
-                flex-direction: column;
-            }
-        }
-
-        .top {
-            gap: 64px;
-
-            @media screen and (max-width: 600px) {
-                gap: 48px;
-            }
-        }
-
-        .middle {
-            gap: 384px;
-
-            @media screen and (max-width: 900px) {
-                gap: 48px;
-            }
-        }
-
-        .bottom {
-            margin-top: -64px;
-
-            @media screen and (max-width: 900px) {
-                justify-content: flex-start;
-                margin-top: 0;
-            }
-        }
-
-        .spell {
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            width: 600px;
-            transform: translate(-50%, -50%);
-            z-index: 0;
-
-            @media screen and (max-width: 1200px) {
-                width: 500px;
-            }
-        }
-
-        .card {
-            position: relative;
-            height: 260px;
-            width: 250px;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: space-around;
-            border: 3px solid var(--primary-color);
-            z-index: 1;
-
-            @media screen and (max-width: 1200px) {
-                height: 210px;
-                width: 200px;
-            }
-
-            @media screen and (max-width: 900px) {
-                flex: 1;
-            }
-
-            @media screen and (max-width: 600px) {
-                width: 100%;
-                flex: initial;
-            }
-
-            .button {
-                bottom: 10%;
-
-                @media screen and (max-width: 1200px) {
-                    min-width: 125px;
-                    padding: 6px;
-                }
-            }
-        }
-
-        .collection {
-            background-image: url('../../assets/pageImages/library2.jpg');
-            background-position: center;
-            background-size: cover;
-        }
-    }
-}
