--- client/src/pages/Collection/Collection.scss
@@ -1,127 +1,0 @@
-.collection {
-    position: relative;
-    align-items: center;
-    background-image: url('../../assets/pageImages/library2.jpg');
-    background-position: center;
-    background-size: cover;
-
-    .container {
-        justify-content: center;
-        flex-direction: row;
-        flex-wrap: wrap;
-
-        .header {
-            width: 50%;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            align-self: flex-start;
-            justify-content: space-around;
-            margin: 64px 25%;
-
-            @media screen and (max-width: 900px) {
-                width: 65%;
-                margin-top: 80px;
-            }
-
-            @media screen and (max-width: 600px) {
-                width: 80%;
-                margin-top: 80px;
-            }
-
-            h1 {
-                width: 300px;
-                align-self: center;
-                padding: 4px 80px;
-                color: var(--white-color);
-                font-size: 48px;
-                font-weight: 700;
-                text-align: center;
-                letter-spacing: 4px;
-
-                @media screen and (max-width: 600px) {
-                    font-size: 28px;
-                    letter-spacing: 2px;
-                }
-            }
-
-            span {
-                margin-bottom: 8px;
-                font-size: 24px;
-                font-weight: 300;
-                letter-spacing: 2px;
-
-                @media screen and (max-width: 600px) {
-                    font-size: 16px;
-                    letter-spacing: 1px;
-                }
-            }
-        }
-
-        .display {
-            width: calc(300px * 0.725);
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            margin: 0 36px;
-
-            @media screen and (max-width: 1200px) {
-                margin: 0 24px;
-            }
-
-            @media screen and (max-width: 900px) {
-                width: calc(200px * 0.725);
-                margin: 0 16px;
-            }
-
-            @media screen and (max-width: 600px) {
-                width: calc(150px * 0.725);
-                margin: 0 6px;
-            }
-
-            .info {
-                height: 100px;
-                font-size: 18px;
-
-                @media screen and (max-width: 900px) {
-                    font-size: 16px;
-                }
-
-                @media screen and (max-width: 600px) {
-                    font-size: 14px;
-                }
-            }
-
-            p {
-                margin: 0;
-                color: var(--secondary-color);
-                text-align: center;
-                text-shadow: 2px 2px 1px var(--accent2-color),
-                    -1px -1px 1px var(--accent2-color),
-                    1px -1px 1px var(--accent2-color),
-                    -1px 1px 1px var(--accent2-color);
-            }
-
-            .disabled {
-                height: 300px;
-                width: calc(300px * 0.725);
-                background-color: var(--black-color);
-                border: 1px solid var(--primary-color);
-                pointer-events: none;
-                filter: blur(1px) grayscale(0.5);
-
-                @media screen and (max-width: 900px) {
-                    height: 200px;
-                    width: calc(200px * 0.725);
-                }
-
-                @media screen and (max-width: 600px) {
-                    height: 150px;
-                    width: calc(150px * 0.725);
-                    border-radius: 8px;
-                }
-            }
-        }
-    }
-}
