--- client/src/pages/Market/Market.scss
@@ -1,1 +1,0 @@
-
