--- client/src/pages/Matchmaking/Matchmaking.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import './Matchmaking.scss'
-
-const Matchmaking = () => {
-    return <div className='matchmaking page'></div>
-}
-
-export default Matchmaking
