--- client/src/pages/Matchmaking/Matchmaking.scss
@@ -1,7 +1,0 @@
-.matchmaking {
-    justify-content: center;
-    background-image: url('../../assets/pageImages/darkportal.jpg');
-    background-position: center;
-    background-size: cover;
-    filter: grayscale(0.7);
-}
