+++ client/src/pages/index.js
@@ -1,13 +1,10 @@
 export { default as Account } from './Account/Account'
-export { default as Arcaneum } from './Arcaneum/Arcaneum'
-export { default as Collection } from './Collection/Collection'
 export { default as Deck } from './Deck/Deck'
 export { default as FirstDeck } from './FirstDeck/FirstDeck'
 export { default as Home } from './Home/Home'
 export { default as LandingPage } from './LandingPage/LandingPage'
 export { default as Match } from './Match/Match'
 export { default as MatchEnd } from './MatchEnd/MatchEnd'
-export { default as Matchmaking } from './Matchmaking/Matchmaking'
 export { default as MatchSetup } from './MatchSetup/MatchSetup'
 export { default as Packs } from './Packs/Packs'
 export { default as Solo } from './Solo/Solo'
